backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.6.0
anthropic==0.64.0
anyio==4.9.0
attrs==25.3.0
bcrypt==4.1.2
beautifulsoup4==4.13.4
black==25.1.0
boto3==1.39.4
botocore==1.39.4
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
choreographer==1.0.9
click==8.2.1
contourpy==1.3.2
cryptography==45.0.5
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
# emergentintegrations==0.1.0
et_xmlfile==2.0.0
Faker==20.1.0
fastapi==0.115.0
filelock==3.18.0
flake8==7.3.0
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.25.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jq==1.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kaleido==1.0.0
kiwisolver==1.4.8
litellm==1.74.2
logistro==1.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
MouseInfo==0.1.3
multidict==6.6.3
mypy==1.16.1
mypy_extensions==1.1.0
narwhals==1.46.0
numpy==2.3.1
oauthlib==3.3.1
openai==1.95.1
openpyxl==3.1.5
orjson==3.10.18
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
passlib==1.7.4
pathspec==0.12.1
pillow==11.3.0
pip==24.0
platformdirs==4.3.8
plotly==6.2.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pyee==13.0.0
pyflakes==3.4.0
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
PyMsgBox==1.0.9
pyotp==2.9.0
pyparsing==3.2.3
pyperclip==1.9.0
pypng==0.20220715.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
python-xlib==0.33
python3-xlib==0.15
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
qrcode==7.4.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
s3transfer==0.13.0
scipy==1.16.0
selenium==4.34.2
setuptools==65.5.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
starlette==0.37.2
stripe==12.3.0
tenacity==8.5.0
tiktoken==0.9.0
tokenizers==0.21.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
typing_extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.30.6
watchfiles==1.1.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wsproto==1.2.0
xlrd==2.0.2
xvfbwrapper==0.2.13
yarl==1.20.1
zipp==3.23.0
